# code/docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "ScopeIQ API",
  "description": "BOM, site design and deployment scoping: documents, discrepancies, redlines, RFIs, BOM revisions, estimates and workflow. Authenticate with POST /auth/login (mock personas).",
  "version": "1.0.0-poc"
 },
 "paths": {
...
  "/sites": {
   "get": {
    "tags": [
     "sites"
    ],
    "summary": "List Sites",
    "operationId": "list_sites_sites_get",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/sites/dashboard": {
   "get": {
    "tags": [
     "sites"
    ],
    "summary": "Dashboard",
    "operationId": "dashboard_sites_dashboard_get",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/sites/{site_id}": {
   "get": {
    "tags": [
     "sites"
    ],
    "summary": "Site Detail",
    "operationId": "site_detail_sites__site_id__get",
    "parameters": [
     {
      "name": "site_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Site Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/sites/{site_id}/config": {
   "get": {
    "tags": [
     "sites"
    ],
    "summary": "Config Lines",
    "operationId": "config_lines_sites__site_id__config_get",
    "parameters": [
     {
      "name": "site_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Site Id"
      }
     },
     {
      "name": "source",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Source"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/sites/{site_id}/evidence": {
   "get": {
    "tags": [
     "sites"
    ],
    "summary": "Evidence",
    "operationId": "evidence_sites__site_id__evidence_get",
    "parameters": [
     {
      "name": "site_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Site Id"
      }
     },
     {
      "name": "selected_only",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Selected Only"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/sites/{site_id}/fields": {
   "get": {
    "tags": [
     "sites"
    ],
    "summary": "Extracted Fields",
    "operationId": "extracted_fields_sites__site_id__fields_get",
    "parameters": [
     {
      "name": "site_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Site Id"
      }
     },
     {
      "name": "needs_review",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "boolean"
        },
        {
         "type": "null"
        }
       ],
       "title": "Needs Review"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/pipeline/sites": {
   "get": {
    "tags": [
     "pipeline"
    ],
    "summary": "Available Sites",
    "operationId": "available_sites_pipeline_sites_get",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/workflow/definition": {
   "get": {
    "tags": [
     "workflow"
    ],
    "summary": "Definition",
    "operationId": "definition_workflow_definition_get",
    "parameters": [
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/workflow/{entity}/{entity_id}/actions": {
   "get": {
    "tags": [
     "workflow"
    ],
    "summary": "Actions",
    "operationId": "actions_workflow__entity___entity_id__actions_get",
    "parameters": [
     {
      "name": "entity",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Entity"
      }
     },
     {
      "name": "entity_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Entity Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/workflow/{entity}/{entity_id}/transition": {
   "post": {
    "tags": [
     "workflow"
    ],
    "summary": "Transition",
    "operationId": "transition_workflow__entity___entity_id__transition_post",
    "parameters": [
     {
      "name": "entity",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Entity"
      }
     },
     {
      "name": "entity_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Entity Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TransitionIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/workflow/{entity}/{entity_id}/history": {
   "get": {
    "tags": [
     "workflow"
    ],
    "summary": "History",
    "operationId": "history_workflow__entity___entity_id__history_get",
    "parameters": [
     {
      "name": "entity",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Entity"
      }
     },
     {
      "name": "entity_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Entity Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/workflow/inbox": {
   "get": {
    "tags": [
     "workflow"
    ],
    "summary": "Inbox",
    "operationId": "inbox_workflow_inbox_get",
    "parameters": [
     {
      "name": "unread_only",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": true,
       "title": "Unread Only"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 100,
       "title": "Limit"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/workflow/inbox/{notification_id}/read": {
   "post": {
    "tags": [
     "workflow"
    ],
    "summary": "Mark Read",
    "operationId": "mark_read_workflow_inbox__notification_id__read_post",
    "parameters": [
     {
      "name": "notification_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Notification Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "TransitionIn": {
    "properties": {
     "action": {
      "type": "string",
      "title": "Action"
     },
     "reason_code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reason Code"
     },
     "comment": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Comment"
     }
    },
    "type": "object",
    "required": [
     "action"
    ],
    "title": "TransitionIn"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}